Move `postgres.py` to `virtool.pg`
* Move connection and table creation to `pg.utils`
* Define `SQLEnum` to extend `Enum` with `to_list()` method
* Move `Base` related code to `pg.base`
